requirements.txt
sentence-transformers
faiss-cpu
PyPDF2
numpy
requests
torch
tqdm
python-dotenv
gradio
